Don't for a new Display when reusing Displays.
This saves the tests from needlessly recreating displays.
When we use a single display for all tests in a process we
should never need to force recreation.

Bug: angleproject:3162

+++ src/tests/test_utils/ANGLETest.cpp
@@ -554,7 +554,7 @@ void ANGLETestBase::ANGLETestSetUp()
     {
         Library *driverLib = ANGLETestEnvironment::GetDriverLibrary(mCurrentParams->driver);
 
-        if (mForceNewDisplay || !mFixture->eglWindow->isDisplayInitialized())
+        if ((mForceNewDisplay && !gReuseDisplays) || !mFixture->eglWindow->isDisplayInitialized())
         {
             mFixture->eglWindow->destroyGL();
             if (!mFixture->eglWindow->initializeDisplay(mFixture->osWindow, driverLib,
@@ -619,7 +619,7 @@ void ANGLETestBase::ANGLETestTearDown()
         checkD3D11SDKLayersMessages();
     }
 
-    if (mFixture->reuseCounter++ >= kWindowReuseLimit || mForceNewDisplay)
+    if (!gReuseDisplays && (mFixture->reuseCounter++ >= kWindowReuseLimit || mForceNewDisplay))
     {
         mFixture->reuseCounter = 0;
         getGLWindow()->destroyGL();
